move job classes in specific modules
