+++ src/osekit/job/config.py
@@ -1,0 +1,36 @@
+from dataclasses import dataclass
+from typing import Literal
+
+from pandas import Timedelta
+
+
+@dataclass
+class JobConfig:
+    """Config used for creating a job.
+
+    Parameters
+    ----------
+    nb_nodes: int
+        Number of nodes on which the job runs.
+    ncpus: int
+        Number of total cores used per node.
+    ngpus: int | None
+        Number of total GPU used per node.
+    mem: str
+        Maximum amount of physical memory used by the job.
+    walltime: str | Timedelta
+        Maximum amount of real time during which the job can be running.
+    venv_name: str
+        Name (or path) of the conda virtual environment in which the job is running.
+    queue: Literal["omp", "mpi"]
+        Queue in which the job will be submitted.
+
+    """
+
+    nb_nodes: int = 1
+    ncpus: int = 2
+    ngpus: int | None = None
+    mem: str = "8gb"
+    walltime: str | Timedelta = "01:00:00"
+    venv_name: str = "osekit"
+    queue: Literal["omp", "mpi"] = "omp"
